refector: 코드 위치 리펙토링
