+++ src/page/community/[postId]/Post.tsx
@@ -10,11 +10,12 @@ import { TextField } from "@common/component/TextField";
 import MoreModal from "@shared/component/MoreModal/MoreModal.tsx";
 import { formatTime } from "@shared/util/formatTime.ts";
 import useModalStore from "@store/moreModalStore.ts";
+
 import {
   useDeleteLike,
   useLikePost,
-} from "@api/domain/community/search/hook.ts";
-import { usePostGet } from "@api/domain/community/post/hook";
+  usePostGet,
+} from "@api/domain/community/post/hook";
 import { useNavigate, useParams } from "react-router-dom";
 import { PATH } from "@route/path.ts";
 import { getAccessToken } from "@api/index.ts";
